requirements.txt
django==4.2
flake8==6.0.0
python-dotenv==1.0
psycopg2-binary==2.9.6
django-split-settings==1.2.0
pytest==7.2.2
django-cors-headers==3.14.0
uWSGI==2.0.21
elasticsearch==8.6.2